+++ schedule-mq-service/src/main/java/com/mixc/cpms/schedule/mq/service/model/TimeBucket.java
@@ -59,26 +59,21 @@ public class TimeBucket {
         list.addAll(msgItemList);
     }
 
-    public List<Long> trigger() {
+    public List<Integer> trigger() {
         long nowSec = TimeKit.nowSeconds();
-
         int size = list.size(), i = 0;
-
         if (log.isDebugEnabled()) {
             log.debug("TimeSegment being trigger tick={} size={} nowSec={}", tick, size, nowSec);
         }
 
-        List<Long> deliverList = new ArrayList<>(32);
-
+        List<Integer> deliverList = new ArrayList<>(32);
         for (Iterator<MsgItem> iterator = list.iterator(); iterator.hasNext() && i <= size; i++) {
 
             MsgItem msg = iterator.next();
             Long deadlineSeconds = msg.getDeadlineSeconds();
-
             if (nowSec < deadlineSeconds) {
                 continue;
             }
-
             deliverList.add(msg.getId());
             iterator.remove();
         }
